Move appbar into a container:recycle:
